[#HOTFIX] CSS improvements

See merge request emundus/cms/core!2520

+++ plugins/fabrik_element/emundus_fileupload/emundus_fileupload.js
@@ -37,10 +37,10 @@ function watch(elementId, attachId) {
 
                 if (result != null) {
                     if (result.limitObtained) {
-                        div.querySelector('button').hide();
+                        div.querySelector('div .btn-upload').hide();
                         div.querySelector('input#'+elementId).hide();
                     } else {
-                        div.querySelector('button').show();
+                        div.querySelector('div .btn-upload').show();
                         div.querySelector('input#'+elementId).show();
                     }
 
@@ -63,28 +63,34 @@ function watch(elementId, attachId) {
                                 divAttachment.appendChild(divLink);
                             }
 
-                            var link = document.createElement('a');
+                            if (result.files[i].can_be_viewed == 1) {
+                                var link = document.createElement('a');
+                                link.setAttribute("href", result.files[i].target);
+                                link.setAttribute("target", "_blank");
+                            } else {
+                                var link = document.createElement('p');
+                            }
                             var linkText = document.createTextNode(result.files[i].local_filename);
-                            link.setAttribute("href", result.files[i].target);
-                            link.setAttribute("target", "_blank");
 
                             divLink.appendChild(link);
                             link.appendChild(linkText);
 
-                            var deleteButton = document.createElement('a');
-                            deleteButton.setAttribute("class", 'em-pointer em-deleteFile em-ml-8');
-                            deleteButton.setAttribute('value', result.files[i].filename);
+                            if (result.files[i].can_be_deleted == 1) {
+                                var deleteButton = document.createElement('a');
+                                deleteButton.setAttribute("class", 'em-pointer em-deleteFile em-ml-8');
+                                deleteButton.setAttribute('value', result.files[i].filename);
 
-                            var deleteIcon = document.createElement('span');
-                            deleteIcon.setAttribute("class", 'material-icons-outlined');
-                            deleteIcon.setAttribute("style",'font-size: 16px');
-                            deleteIcon.appendChild(document.createTextNode('clear'));
-                            deleteButton.appendChild(deleteIcon);
+                                var deleteIcon = document.createElement('span');
+                                deleteIcon.setAttribute("class", 'material-icons-outlined');
+                                deleteIcon.setAttribute("style",'font-size: 16px');
+                                deleteIcon.appendChild(document.createTextNode('clear'));
+                                deleteButton.appendChild(deleteIcon);
 
-                            divLink.appendChild(deleteButton);
+                                divLink.appendChild(deleteButton);
 
-                            var button = document.querySelector('#' + elementId + '_attachment_link' + i + ' > a.em-deleteFile');
-                            button.addEventListener('click', () => FbFileUpload.delete(elementId, attachId));
+                                var button = document.querySelector('#' + elementId + '_attachment_link' + i + ' > a.em-deleteFile');
+                                button.addEventListener('click', () => FbFileUpload.delete(elementId, attachId));
+                            }
                         }
                     }
 
@@ -300,7 +306,9 @@ var FbFileUpload = {
                         });
 
                         input.value = '';
-                        deleteButton.style.display = 'none';
+                        if(deleteButton != null) {
+                            deleteButton.style.display = 'none';
+                        }
                     }
 
                     if (result[j].encrypt == false){
@@ -404,7 +412,7 @@ var FbFileUpload = {
                     }
                 }).then((res) => {
                     if (res.status) {
-                        div_parent.querySelector('button').show();
+                        div_parent.querySelector('div.btn-upload').show();
                         div_parent.querySelector('input#'+elementId).show();
 
                         file.parentElement.parentElement.remove();
